Retrieve ovn-k upstream issues and create jira stories tracking them
Ovn-kubernetes upstream issues labeled with "ci-flake" are not tracked in JIRA: let's create a jira user story under the SDN-4175 epic for each such open issue:
- the assignee is kept whenever possible, otherwise bbennett is used.
- the summary (title) of these jira cards will start with upstream-$GITHUB_ISSUE_ID, where $GITHUB_ISSUE_ID is the ID found in the URL of the issue itself (e.g. upstream-3986 for ovn-org/ovn-kubernetes#3986)

A list of which jira stories have been created is printed to stdout, as well as a list of stories in the SDN-4175 epic whose status doesn't match the status of the github issue they're tracking.

The new "-g" ("--process-github-issues") input parameter executes this new functionality, or the script alone without any input parameters.

Signed-off-by: Riccardo Ravaioli <rravaiol@redhat.com>