Align status and assignee between jira and github
Github is the source of truth. For a given github issue, the jira story that tracks it must show the same status (open / closed) and the same assignee; if not, script will align it. If ever more than one jira story tracks the same github issue, then the jira story with the lowest ID is kept and all other ones are closed.

Signed-off-by: Riccardo Ravaioli <rravaiol@redhat.com>