Add tooling to reduce bug backlog toil for viewing bz comment updates
.. for the last 5 days..

The problem I am trying to solve is that there is
not enough resources to manage the bugs on the bot,
so we need some tooling to help.
This helps give our networking consumers confidence
we are listening and importantly reacting to their
updates.

New tool idea in future might also print bugs with severity
or priorty change.

Tool will print out bugs that have new comments within the
last 4 days with a URL you can click on.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>