Bug backlog tooling: add arg for listing stale bugs (>4 months)
We are currently keeping around a lot of bugs on the
bot that we are never going to work on.
It causes huge cognitive load to manage for a single
bug dispatch manager.
Its even worse than that because we should probably
just document a lot of the issues and thats not
getting done when the bugs are sitting on the bot.
We have a lot of high value efforts going on and
managing low value bugs is key for us to scale
as a team. Lets add this tool and others to lower
the toil.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>